Implement JobDataService v 0.1, update interfaces, update shared Job

+++ Jobtracker.shared/src/main/java/net/greenbeansit/jobtracker/shared/Job.java
@@ -51,7 +51,7 @@ public class Job implements Serializable
 		this.jobID = jobID;
 	}
 
-	public long getMaxBudget()
+	public Integer getMaxBudget()
 	{
 		return maxBudget;
 	}
